+++ src/ML/Engine/Preferences.cpp
@@ -62,10 +62,10 @@ namespace ml
 		);
 	}
 
-	float Preferences::GetDouble(const String & section, const String & name, double default_value) const
+	double Preferences::GetDouble(const String & section, const String & name, double default_value) const
 	{
 		return ((m_ini)
-			? (float)(static_cast<INIReader *>(m_ini)->GetReal(
+			? (static_cast<INIReader *>(m_ini)->GetReal(
 				section,
 				name,
 				default_value
